Adds Add page and correctly adds user id to entry and only logged in users can add entries
